Updated while loop
Correction was made to somebools.lenght. Arrays are 0 base index so
someBools.length = 8, in order get 7 you must subtract 1 to the length.
Another option would be to create a variable that holds the number 7.
Because of the pattern, a simpler form would be to cut the array in
half by dividing the .length by 2.